Use /dev/fd/ instead of the original complicated construct.
This is just a simplification of the for checking number of opened (and
closed) file descriptors.

Overcomplicated construct

    self._proc = "/proc/{0}/fd/".format(os.getpid())

should be at least simplified to

    self._proc = "/proc/self/fd/"

(so that we don't needlesly call Python function to get PID etc. for
nothing). However, when we are there already, there is nothing more
simple than to use symlink (on Linux) /dev/fd -> /proc/self/fd which
also happens to be compatible on all UNIX systems.

For those who are still fearful, Documentation/admin-guide/devices.rst
is the official documentation and it claims that this symlink is
compulsory, so it must be on all Linuces.